optimize

+++ src/main/java/cn/xiaozi0721/bpk/interfaces/IPlayerResizable.java
@@ -1,7 +1,7 @@
 package cn.xiaozi0721.bpk.interfaces;
 
 public interface IPlayerResizable {
-    double BPKMod$getSneakHeight();
+    float BPKMod$getSneakHeight();
     boolean BPKMod$getUnderBlock();
     boolean BPKMod$getResizingAllowed();
 }
